ml_models/pyproject.toml
 [build-system]
 requires = ["setuptools>=65", "wheel"]
 build-backend = "setuptools.build_meta"

 [project]
 name = "bushfire-ai"
 version = "0.1.0"
 description = "Real-time fire, bushfire, and smoke detection pipeline for 5G-enabled drones"
 readme = "README.md"
 requires-python = ">=3.10"
 license = {text = "MIT"}
 authors = [
   {name = "Bushfire Hackathon Team"}
 ]
 classifiers = [
   "Programming Language :: Python :: 3",
   "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent",
   "Topic :: Scientific/Engineering :: Artificial Intelligence"
 ]

 dependencies = [
   "opencv-python>=4.9.0",
   "torch>=2.2.0",
   "torchvision>=0.17.0",
   "ultralytics>=8.1.0",
   "pydantic>=2.6",
   "requests>=2.31",
   "httpx>=0.27",
   "python-dotenv>=1.0",
   "loguru>=0.7"
 ]

 [project.optional-dependencies]
 dev = [
   "pytest>=8.0",
   "pytest-asyncio>=0.23",
   "pre-commit>=3.5",
   "ruff>=0.3"
 ]

 [tool.setuptools.packages.find]
 where = ["src"]

